Removed some dead code
Removed dead code form the Wagon class.

Made a beginning for the Unit Test of Scenario 3 since this one still doesn't work like intended. The only possible outcome should be 4 but 3 out of 4 of my sorting methods say that they can do it with 3 wagons. So they add the small carnivore to the wagon with the big-, medium- and small herbivores.